requirements.txt
spacy
openai-whisper
argostranslate
pytube
